+++ lib/src/common/repos_service.dart
@@ -957,6 +957,18 @@ class RepositoriesService extends Service {
     );
   }
 
+  /// Lists the latest release for the specified repository.
+  ///
+  /// API docs: https://developer.github.com/v3/repos/releases/#get-the-latest-release
+  Future<Release> getLatestRelease(RepositorySlug slug) {
+    ArgumentError.checkNotNull(slug);
+    return github.getJSON<Map<String, dynamic>, Release>(
+      '/repos/${slug.fullName}/releases/latest',
+      convert: (i) => Release.fromJson(i),
+      statusCode: StatusCodes.OK,
+    );
+  }
+
   /// Fetches a single release by the release ID.
   ///
   /// API docs: https://developer.github.com/v3/repos/releases/#get-a-single-release
@@ -972,13 +984,23 @@ class RepositoriesService extends Service {
 
   /// Fetches a single release by the release tag name.
   ///
+  /// Throws a [ReleaseNotFound] exception if the release
+  /// doesn't exist.
+  ///
   /// API docs: https://developer.github.com/v3/repos/releases/#get-a-release-by-tag-name
-  Future<Release> getReleaseByTagName(RepositorySlug slug, String tagName) =>
-      github.getJSON(
-        '/repos/${slug.fullName}/releases/tags/$tagName',
-        convert: (i) => Release.fromJson(i),
-        statusCode: StatusCodes.OK,
-      );
+  Future<Release> getReleaseByTagName(
+      RepositorySlug slug, String tagName) async {
+    return github.getJSON(
+      '/repos/${slug.fullName}/releases/tags/$tagName',
+      convert: (i) => Release.fromJson(i),
+      statusCode: StatusCodes.OK,
+      fail: (http.Response response) {
+        if (response.statusCode == 404) {
+          throw ReleaseNotFound.fromTagName(github, tagName);
+        }
+      },
+    );
+  }
 
   /// Creates a Release based on the specified [createRelease].
   ///